requirements.txt
Flask==2.2.5
requests==2.28.2
Werkzeug==2.2.3
MarkupSafe==2.1.2
spotipy==2.22.1
mysql-connector-python==8.4.0